requirements.txt
pdoc==12.*
flake8==6.0.0
pytest==7.1.*
pandas==2.0.1
numpy==1.23.5
imutils==0.5.4
adabelief-tf==0.2.1
mlflow==2.3.2
Pillow==9.5.0
opencv-python==4.7.0.72
scikit-image==0.20.0
tensorflow==2.12.3